requirements.txt
beautifulsoup4
flask
imdbpy
